Set up Level scene - click to move Supergirl
Create initial level scene. Add background sprite and supergirl (Hero)
sprite.

Clicks on the background sprite send commands to the hero sprite telling
it where to fly next. the hero maintains a queue of flight destinations
in case it can not keep up with the rate of clicking. Commands are
designed to support including which minion to pick up and which note to
place it on, but this is not yet implemented.